+++ R/guide-colorbar.R
@@ -15,45 +15,18 @@ NULL
 #' see [guides()].
 #'
 #' @inheritParams guide_legend
-#' @param barwidth,barheight A numeric or [grid::unit()] object specifying the
-#'   width and height of the bar respectively. Default value is derived from
-#'   `legend.key.width`, `legend.key.height` or `legend.key` in [theme()].\cr
-#'   `r lifecycle::badge("experimental")`: optionally a `"null"` unit to stretch
-#'   the bar to the available space.
-#' @param frame A theme object for rendering a frame drawn around the bar.
-#'   Usually, the object of `element_rect()` is expected. If `element_blank()`
-#'   (default), no frame is drawn.
-#' @param frame.colour A string specifying the colour of the frame
-#'   drawn around the bar. For backward compatibility, if this argument is
-#'   not `NULL`, the `frame` argument will be set to `element_rect()`.
-#' @param frame.linewidth A numeric specifying the width of the frame
-#'   drawn around the bar in millimetres.
-#' @param frame.linetype A numeric specifying the linetype of the frame
-#'   drawn around the bar.
 #' @param nbin A numeric specifying the number of bins for drawing the
 #'   colourbar. A smoother colourbar results from a larger value.
 #' @param raster A logical. If `TRUE` then the colourbar is rendered as a
 #'   raster object. If `FALSE` then the colourbar is rendered as a set of
 #'   rectangles. Note that not all graphics devices are capable of rendering
 #'   raster image.
-#' @param ticks A theme object for rendering tick marks at the colourbar.
-#'   Usually, the object of `element_line()` is expected (default). If
-#'   `element_blank()`, no tick marks are drawn. For backward compatibility,
-#'   can also be a logical which translates `TRUE` to `element_line()` and
-#'   `FALSE` to `element_blank()`.
-#' @param ticks.colour A string specifying the colour of the tick marks.
-#' @param ticks.linewidth A numeric specifying the width of the tick marks in
-#'   millimetres.
-#' @param ticks.length A numeric or a [grid::unit()] object specifying the
-#'   length of tick marks at the colourbar.
 #' @param draw.ulim A logical specifying if the upper limit tick marks should
 #'   be visible.
 #' @param draw.llim A logical specifying if the lower limit tick marks should
 #'   be visible.
 #' @param direction  A character string indicating the direction of the guide.
 #'   One of "horizontal" or "vertical."
-#' @param default.unit A character string indicating [grid::unit()]
-#'   for `barwidth` and `barheight`.
 #' @param reverse logical. If `TRUE` the colourbar is reversed. By default,
 #'   the highest value is on the top and the lowest value is on the bottom
 #' @param available_aes A vector of character strings listing the aesthetics
@@ -77,19 +50,31 @@ NULL
 #' # Control styles
 #'
 #' # bar size
-#' p1 + guides(fill = guide_colourbar(barwidth = 0.5, barheight = 10))
+#' p1 + guides(fill = guide_colourbar(theme = theme(
+#'   legend.key.width  = unit(0.5, "lines"),
+#'   legend.key.height = unit(10, "lines")
+#' )))
+#'
 #'
 #' # no label
-#' p1 + guides(fill = guide_colourbar(label = FALSE))
+#' p1 + guides(fill = guide_colourbar(theme = theme(
+#'   legend.text = element_blank()
+#' )))
 #'
 #' # no tick marks
-#' p1 + guides(fill = guide_colourbar(ticks = FALSE))
+#' p1 + guides(fill = guide_colourbar(theme = theme(
+#'   legend.ticks = element_blank()
+#' )))
 #'
 #' # label position
-#' p1 + guides(fill = guide_colourbar(label.position = "left"))
+#' p1 + guides(fill = guide_colourbar(theme = theme(
+#'   legend.text.position = "left"
+#' )))
 #'
 #' # label theme
-#' p1 + guides(fill = guide_colourbar(label.theme = element_text(colour = "blue", angle = 0)))
+#' p1 + guides(fill = guide_colourbar(theme = theme(
+#'   legend.text = element_text(colour = "blue", angle = 0)
+#' )))
 #'
 #' # small number of bins
 #' p1 + guides(fill = guide_colourbar(nbin = 3))
@@ -102,7 +87,7 @@ NULL
 #'   scale_fill_continuous(
 #'     limits = c(0,20), breaks = c(0, 5, 10, 15, 20),
 #'     guide = guide_colourbar(nbin = 100, draw.ulim = FALSE, draw.llim = FALSE)
-#'    )
+#'   )
 #'
 #' # guides can be controlled independently
 #' p2 +
@@ -111,148 +96,45 @@ NULL
 #' p2 + guides(fill = "colourbar", size = "legend")
 #'
 #' p2 +
-#'   scale_fill_continuous(guide = guide_colourbar(direction = "horizontal")) +
-#'   scale_size(guide = guide_legend(direction = "vertical"))
+#'   scale_fill_continuous(guide = guide_colourbar(theme = theme(
+#'     legend.direction = "horizontal"
+#'   ))) +
+#'   scale_size(guide = guide_legend(theme = theme(
+#'     legend.direction = "vertical"
+#'   )))
 guide_colourbar <- function(
-
-  # title
   title = waiver(),
-  title.position = NULL,
-  title.theme = NULL,
-  title.hjust = NULL,
-  title.vjust = NULL,
-
-  # label
-  label = TRUE,
-  label.position = NULL,
-  label.theme = NULL,
-  label.hjust = NULL,
-  label.vjust = NULL,
-
-  # bar
-  barwidth = NULL,
-  barheight = NULL,
+  theme = NULL,
   nbin = 300,
   raster = TRUE,
-
-  # frame
-  frame = element_blank(),
-  frame.colour = NULL,
-  frame.linewidth = NULL,
-  frame.linetype = NULL,
-
-  # ticks
-  ticks = element_line(),
-  ticks.colour = NULL,
-  ticks.linewidth = NULL,
-  ticks.length = unit(0.2, "npc"),
   draw.ulim = TRUE,
   draw.llim = TRUE,
-
-  # general
   position = NULL,
   direction = NULL,
-  default.unit = "line",
   reverse = FALSE,
   order = 0,
   available_aes = c("colour", "color", "fill"),
   ...
 ) {
-  if (!(is.null(barwidth) || is.unit(barwidth))) {
-    barwidth <- unit(barwidth, default.unit)
-  }
-  if (!(is.null(barheight) || is.unit(barheight))) {
-    barheight <- unit(barheight, default.unit)
-  }
-  if (!is.unit(ticks.length)) {
-    ticks.length <- unit(ticks.length, default.unit)
-  }
 
-  if (!is.null(title.position)) {
-    title.position <- arg_match0(title.position, .trbl)
-  }
+  theme <- deprecated_guide_args(theme, ...)
   if (!is.null(position)) {
     position <- arg_match0(position, c(.trbl, "inside"))
   }
-  if (!is.null(direction)) {
-    direction <- arg_match0(direction, c("horizontal", "vertical"))
-  }
-  if (!is.null(label.position)) {
-    label.position <- arg_match0(label.position, .trbl)
-  }
-
-  if (!is.null(frame.colour) && !inherits(frame, "element_rect")) {
-    # For backward compatibility, frame should not be element_blank when
-    # colour is not NULL
-    cli::cli_inform(c(paste0(
-      "If {.arg frame.colour} is set, {.arg frame} should not be ",
-      "{.cls {class(frame)[[1]]}}."
-    ), "i" = "{.arg frame} has been converted to {.cls element_rect}."))
-    frame <- element_rect()
-  }
-  if (inherits(frame, "element_rect")) {
-    frame$colour    <- frame.colour    %||% frame$colour
-    frame$linewidth <- frame.linewidth %||% frame$linewidth %||% (0.5 / .pt)
-    frame$linetype  <- frame.linetype  %||% frame$linetype  %||% 1
-  } else {
-    frame <- element_blank()
-  }
-
-  if (is.logical(ticks)) {
-    # Also for backward compatibility. `ticks = FALSE` used to mean: don't draw
-    # the ticks
-    ticks <- if (ticks) element_line() else element_blank()
-  }
-  if (inherits(ticks, "element_line")) {
-    ticks$colour    <- ticks.colour    %||% ticks$colour    %||% "white"
-    ticks$linewidth <- ticks.linewidth %||% ticks$linewidth %||% (0.5 / .pt)
-  }
-
-  # Trick to re-use this constructor in `guide_coloursteps()`.
-  args  <- list2(...)
-  super <- args$super %||% GuideColourbar
-  args$super <- NULL
 
   new_guide(
-    # title
     title = title,
-    title.position = title.position,
-    title.theme = title.theme,
-    title.hjust = title.hjust,
-    title.vjust = title.vjust,
-
-    # label
-    label = label,
-    label.position = label.position,
-    label.theme = label.theme,
-    label.hjust = label.hjust,
-    label.vjust = label.vjust,
-
-    # bar
-    keywidth = barwidth,
-    keyheight = barheight,
+    theme = theme,
     nbin = nbin,
     raster = raster,
-
-    # frame
-    frame = frame,
-
-    # ticks
-    ticks = ticks,
-    ticks_length = ticks.length,
     draw_lim = c(isTRUE(draw.llim), isTRUE(draw.ulim)),
-
-    # general
     position = position,
     direction = direction,
     reverse = reverse,
     order = order,
-
-    # parameter
     available_aes = available_aes,
     name = "colourbar",
-    !!!args,
-    super = super
+    super = GuideColourbar
   )
 }
 
@@ -270,21 +152,14 @@ GuideColourbar <- ggproto(
   params = list(
     # title
     title = waiver(),
-    title.position = NULL,
-    title.theme = NULL,
-    title.hjust = NULL,
-    title.vjust = NULL,
 
-    # label
-    label = TRUE,
-    label.position = NULL,
-    label.theme = NULL,
-    label.hjust = NULL,
-    label.vjust = NULL,
+    # theming
+    theme = NULL,
+    default_ticks = element_line(colour = "white", linewidth = 0.5 / .pt),
+    default_frame = element_blank(),
+    default_tick_length = unit(0.2, "npc"),
 
     # bar
-    keywidth  = NULL,
-    keyheight = NULL,
     nbin = 300,
     raster = TRUE,
 
@@ -306,16 +181,19 @@ GuideColourbar <- ggproto(
   hashables = exprs(title, key$.label, decor, name),
 
   elements = list(
-    frame       = "rect",
-    ticks       = "line",
-    ticks_length = unit(0.2, "npc"),
-    background  = "legend.background",
-    margin      = "legend.margin",
-    key         = "legend.key",
-    key.height  = "legend.key.height",
-    key.width   = "legend.key.width",
-    text        = "legend.text",
-    theme.title = "legend.title"
+    background     = "legend.background",
+    margin         = "legend.margin",
+    key            = "legend.key",
+    key_height     = "legend.key.height",
+    key_width      = "legend.key.width",
+    text           = "legend.text",
+    theme.title    = "legend.title",
+    text_position  = "legend.text.position",
+    title_position = "legend.title.position",
+    axis_line      = "legend.axis.line",
+    ticks          = "legend.ticks",
+    ticks_length   = "legend.ticks.length",
+    frame          = "legend.frame"
   ),
 
   extract_key = function(scale, aesthetic, ...) {
@@ -346,11 +224,9 @@ GuideColourbar <- ggproto(
 
   extract_params = function(scale, params,
                             title  = waiver(), ...) {
-    params$title <- scale$make_title(
-      params$title %|W|% scale$name %|W|% title
-    )
+    params$title <- scale$make_title(params$title %|W|% scale$name %|W|% title)
 
-    limits <- c(params$decor$value[1], params$decor$value[nrow(params$decor)])
+    limits <- params$decor$value[c(1L, nrow(params$decor))]
     params$key$.value <- rescale(
       params$key$.value,
       c(0.5, params$nbin - 0.5) / params$nbin,
@@ -374,38 +250,41 @@ GuideColourbar <- ggproto(
       params$direction,
       c("horizontal", "vertical"), arg_nm = "direction"
     )
-    valid_label_pos <- switch(
-      params$direction,
-      "horizontal" = c("bottom", "top"),
-      "vertical"   = c("right", "left")
-    )
-    params$label.position <- params$label.position %||% valid_label_pos[1]
-    if (!params$label.position %in% valid_label_pos) {
-      cli::cli_abort(paste0(
-        "When {.arg direction} is {.val {params$direction}}, ",
-        "{.arg label.position} must be one of {.or {.val {valid_label_pos}}}, ",
-        "not {.val {params$label.position}}."
-      ))
-    }
-    params$title.position <- arg_match0(
-      params$title.position %||%
-        switch(params$direction, vertical = "top", horizontal = "left"),
-      .trbl, arg_nm = "title.position"
-    )
-    params$rejust_labels <- FALSE
     params
   },
 
-  override_elements = function(params, elements, theme) {
-    # These key sizes are the defaults, the GuideLegend method may overrule this
+  setup_elements = function(params, elements, theme) {
+    # We set the defaults in `theme` so that the `params$theme` can still
+    # overrule defaults given here
     if (params$direction == "horizontal") {
-      elements$key.width <- elements$key.width * 5
+      theme$legend.key.width  <- theme$legend.key.width * 5
+      valid_position <- c("bottom", "top")
     } else {
-      elements$key.height <- elements$key.height * 5
+      theme$legend.key.height <- theme$legend.key.height * 5
+      valid_position <- c("right", "left")
     }
-    elements$ticks <- combine_elements(elements$ticks, theme$line)
-    elements$frame <- combine_elements(elements$frame, theme$rect)
-    GuideLegend$override_elements(params, elements, theme)
+
+    # Set defaults
+    theme <- replace_null(
+      theme,
+      legend.text.position = valid_position[1],
+      legend.ticks.length  = params$default_tick_length,
+      legend.ticks         = params$default_ticks,
+      legend.frame         = params$default_frame
+    )
+
+    # Let the legend guide handle the rest
+    elements <- GuideLegend$setup_elements(params, elements, theme)
+
+    # Check text position
+    if (!elements$text_position %in% valid_position) {
+      cli::cli_abort(paste0(
+        "When {.arg direction} is {.val {params$direction}}, ",
+        "{.arg legend.text.position} must be one of ",
+        "{.or {.val {valid_position}}}, not {.val {elements$text_position}}."
+      ))
+    }
+    elements
   },
 
   build_labels = function(key, elements, params) {
@@ -485,8 +364,8 @@ GuideColourbar <- ggproto(
 
   measure_grobs = function(grobs, params, elements) {
     params$sizes <- list(
-      widths  = elements$key.width,
-      heights = elements$key.height
+      widths  = elements$width_cm,
+      heights = elements$height_cm
     )
     GuideLegend$measure_grobs(grobs, params, elements)
   }